chore(boundary): `ic-gateway` uvm (#5303)
The goal is to substitute the
[BoundaryNode](https://github.com/dfinity/ic/blob/3c10bb9cc1570b9549b8ab6bf0fa29e038048c37/rs/tests/driver/src/driver/boundary_node.rs#L84)
component used in `system-tests`/`testnets` with the current mainnet
architecture, IC HTTP Gateway -> API boundary nodes -> Replica nodes.

The scope of this PR is to:
- Create a minimalistic `IcGatewayVm` - a wrapper around Farm's
Universal VM (UVM). This UVM runs `ic-gateway` service in a docker,
which serves IC traffic through existing API boundary nodes.
- Substitute usage of `BoundaryNode` in `system-tests` with a
combination of `IcGatewayVm` and API boundary node/s.
- Substitute `BoundaryNode` with `IcGatewayVm` in all testnets
- Add `ic-gateway` scraping target to `PrometheusVm`