feat(nns): Support disbursing maturity to an AccountIdentifier (#5351)
# Why

Currently, `DisburseMaturity` supports disbursing into any icrc1
account. However, we still have valid use cases beyond icrc1 account,
for example for disbursing into a CEX account, which typically does not
use icrc1 account but a 32-byte account identifier.

# What

* Add `to_account_identifier` as a new field in the `DisburseMaturity`
command input
* Change `account_to_disbures_to` in the storage type into a `oneof
destination` with the previous `account_to_disbures_to` as one variant.
* Implement `Destination::try_new` to handle the 3 cases - (a) no
account or account identifier (use caller), (b) account, and (c) account
identifier
* Implement the conversion from `Destination` to `AccountIdentifier` for
the finalization (for minting ICPs).
* Implement the conversion from `Destination` to `Option<Account>` and
`Option<AccountIdentifier>` for the fields in "full neuron"
* Perform a similar validation for neuron management proposals


# Backward Compatibility

The main concern of the backward compatibility is that there can already
be disbursements prior to this PR, and they do not use the `oneof
destination` but the previous schema with only `account_to_disbures_to`.
The `oneof` is designed to be compatible with the previous schema, and
the `test_deserialize_maturity_disbursement_from_old_format` test is
added to make sure that the decoding actually works, at the `Storable`
level.