chore(nns): Delete //rs/nns/governance:scale_bench (#5397)
# Why

The benchmark target has been broken for a long time and CI does not run
it. This target focuses on benchmarking the proposal making process with
different number of voting eligible neurons. The computational intensive
part is to create neurons, and such benchmark already exists with
canbench (compute_ballots_for_new_proposal_with_stable_neurons).
Therefore, it should be OK to delete it.