fix: increase timeout for the {upgrade,downgrade}_app_subnet_tests (#5436)
The `//rs/tests/consensus/upgrade:{upgrade,downgrade}_app_subnet_tests`
are highly flaky because they timeout frequently. When looking at their
Kibana logs we see that the node was in the process of rebooting to the
upgraded version but then the test timed out. So giving it a bit more
time will probably allow these tests to succeed. So we increase their
bazel timeouts from `long` (15 minutes) to `eternal` (1 hour).