refactor(nns): Refactor Governance::new to combine conversion and fallback (#5419)
# Why

* Before this PR, the `Governance::new` first converts the governance
API type (including neurons) to the internal type (including neurons),
then the neurons are split out of the internal type when converting to
`HeapGovernanceData`. During this process, the
`GovernanceProto::neurons` field is used temporarily, this field is
always empty in production, and we want to clean up this field.
* The 2 step conversion is also more complicated than it could be, and
for many fields the direct conversion becomes simpler

# What

* Combine the effect of `api::Governance->GovernanceProto`, "fill
missing fields with defaults", and `split_governance_proto` into one
function
* Instead of convert from API->protobuf->internal for conversion rate,
convert API->internal directly (with fallback).