feat(sns): Implement SnsGov.get_metrics (#5185)
In this PR, the internal functionality to interact with ledger canister
and fetch the last ledger, as well as counting the governance proposals
in a given time window, is implemented. By merging this PR, a new
function `get_metrics()` would be exposed on the governance canister. By
calling this function, users can get recent activity metrics of the SNS
governance during their time period of interest.

< [Previous PR](https://github.com/dfinity/ic/pull/5079) | [Next
PR](https://github.com/dfinity/ic/pull/5374) >

---------

Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>
Co-authored-by: Marko Kosmerl <marko.kosmerl@dfinity.org>