test(nns): Remove (prop_)assert_changes (#5421)
# Why

The ability to assert on the entirety of the NNS Governance state was OK
before, but as we have more and more data in stable storage, it becomes
less feasible to do so. In particular now neurons live in the stable
storage, so we won't be able assert on the neurons anyway.

# What

* Remove (prop_)assert_changes and its invocations
* For `check_proposal_status_after_voting_and_after_expiration_new`,
there is already a test
`check_proposal_status_after_voting_and_after_expiration` which does the
same thing except for asserting the global state changes.