chore(nns): Remove GovernanceProto::neurons (#5442)
# Why

The GovernanceProto::neurons are always empty in production (since
neurons are in stable memory). To see this, we can look at the
[split|reassemble]_governance_proto methods and see that the neurons are
discarded when splitting, and BTreeMap::new() is used when reassembling,
and it has been the case since a month ago (#5007). Therefore, dropping
the field should be safe.

# What

* Drop and reserve the `neurons` field in governance.proto
* Remove the usage of the field
* Removing the neuron count in the post_upgrade print (it has been
showing "neuron count: 0" anyway)