refactor(nns): Move Followees out of Neuron in governance.proto (#5456)
# Why

To prepare for the change to entire remove the neuron prost type (#5460)
while minimizing a buf-breaking change, this PR is sent with only the
buf-breaking change.

# What

* Move `Followees` outside of the `Neuron` definition
* Generate protobuf
* Change all references `neuron::Followees` -> `Followees`

# Backward compatibility

The `Neuron.Followees` and `Followees` types should have the same wire
format and hence decoding an old value should be fine. In addition, the
Governance canister should no longer have data with this field. Just to
be more certain, a draft PR (#5461) shows that a `Governance` with or
without default_followees can be decoded from a message encoded with the
previous version.