chore: CON-1504 Increase acceptable gap between next pending CUP height and notarized height to 130 (#5448)
On subnets with large state, manifest computation may take a long time,
which delays the creation of the CUP. If it takes long enough, the
subnet may exceed the acceptable gap between the pending CUP height and
the latest notarized height. In this case, the subnet stalls until the
manifest and CUP are created, before it continues notarizing new blocks.

As a short term solution, this PR increases the acceptable gap from 70
to 130 heights. In order to limit the increase on our bound of the
maximum number of artifacts, we _keep_ the acceptable gap between
notarized height and _certified_ height at 70. This requires us to
slightly alter the logic for computing the bouncer value, and the
maximum pool bounds. Before, these functions implicitly assumed that
both bounds would be equal.