chore(IDX): remove zh1 pin from daily jobs (#5476)
All CI jobs succeeded, except for `Bazel Test NNS Nightly` which first
timed out, then failed when I bumped the timeout up to 30min. However,
it has also been failing on regular runs, so it's difficult to identify
if this is related to it being run in dm1.

See failed run here:
https://dash.dm1-idx1.dfinity.network/invocation/a1d7914e-e0ee-4657-b492-39383df7b73a

---------

Co-authored-by: IDX GitHub Automation <>