feat(node): guestos-recovery-engine component (#5355)
NODE-1599

The guestos-recovery-engine component will be used in the event of an
NNS recovery.

First, the guestos-recovery- **upgrader** component will be used to
manually upgrade the GuestOS to a guestos-recovery version that contains
the guestos-recovery**-engine** component.

The guestos-recovery-engine component then completes the recovery
process by downloading and applying the recovery artifacts

I've created recovery GuestOS variants that will only be run manually
that include the guestos-recovery-engine

---------

Co-authored-by: Eero Kelly <eero.kelly@dfinity.org>