feat(registry): Enable chunking. (#5210)
This is essentially "just" a 1-line change, but it activates a fair
amount of new code and behavior that was previously dormant in the
Registry canister. To trace the code being activated, I suggest that the
reader start with [this query][trace].

[trace]:
https://sourcegraph.com/search?q=repo:%5Egithub%5C.com/dfinity/ic%24+is_chunkifying_large_values_enabled%5C%28&patternType=regexp&case=yes&sm=0

This is a breaking change, but clients have been on notice since [this
announcement at the end of March][announce]. There, instructions on how
to migrate were provided. To briefly summarize the client migration
procedure, when reading a large mutation or record, clients must make
follow up calls to a new `get_chunk` method. In short, this essentially
adds pagination.

[announce]:
https://forum.dfinity.org/t/breaking-registry-changes-for-large-records/42893/5?u=daniel-wong

Clients in the `ic` repo have already been migrated. The main thing
blocking this PR is for those clients to be re-built and re-deployed.

# When Clients Ready?

To figure out who needs to migrate, I ran this:

```
bazel query 'kind(
  .*bin.*,
  rdeps(//..., //rs/registry/transport, 999999)
)' \
| grep -v ':_' \
| grep -v test \
| grep -v bench
```

Then, I manually cross referenced with `CODEOWNERS`, and went around and
confirmed that recent enough builds have been deployed.

# References

Closes [NNS1-3803].

[NNS1-3803]: https://dfinity.atlassian.net/browse/NNS1-3803