fix(IDX): use different zig cache for configs (#5531)
This works around a [bug](https://github.com/ziglang/zig/issues/23993)
in the zig cache. The bug is triggered when bazel tries to build several
"almost" identical targets at the same time where only the config
differs.

This is the case for us with jemalloc, where transitions set different
_rustc_ flags for the build, but which bazel still interprets as
different configs.

This causes concurrent builds of jemalloc and trips the zig cache.