chore: require build determinism for all targets except a few exclusions (#5539)
The goal is to have as few non-reproducible bazel targets as possible
because non-reproducible targets hurt cache utilisation and increase the
risk of having non-reproducible final artifacts like guesOS images.

So we forbid all non-reproducible targets except for a few known ones
which we can document and ideally link to an upstream ticket.