fix(registry): Handle currently impossible edge case in canister_ranges_ updates (#5545)
This adds a test, and fixes the edge case where a new routing table
could have a mutation that doesn't have an existing range to fall into.
This is currently not possible, but this change makes the code more
robust so that it cannot unnecessarily panic if there's no range
starting at CanisterId "0".