fix: Consolidate how caller is extracted from System API (#5542)
This PR consolidates how caller is extracted from System API, forcing to
go through `ApiType::caller` for all cases. There used to be a second
method in `SystemApiImpl` which would actually return a slightly
different result than `ApiType::caller` (a miss in an older change
https://github.com/dfinity/ic/commit/e2ea0f12e3e4465d3dccd44d1e4988fc499dc2d0
that adjusted when `ic0.msg_caller*` is available).

With this fix consistency is ensured when the caller needs to be
extracted from the System API and `ApiType::caller` is covered now
through our usual system api availability tests for future proofing.