refactor: [EXC-2052] Use CompositeQuery System Api type (#5550)
This PR is the first step to align the system API implementation with
the interface spec as far as the system api types are concerned. It
introduces a `CompositeQuery` type that essentially decouples it from
`NonReplicatedQuery` and the extra `NonReplicatedQueryKind` that was
used in the system API can be dropped. In a follow-up, the composite
reply and reject callbacks will be added to decouple them from the
replicated reply and reject callbacks.