chore(IDX): add alerts for ic-private (#5554)
This adds an alert for `Sync IC private from IC public` so we will be
notified if it is broken (like it was last week).

This workflow will be triggered from `ic-private`, though because all
workflows except for `Sync IC private from IC public` are disabled we
should only receive alerts for this one.

If this alert fails and we start getting alerts every 10 minutes, we can
always disable this `Slack Workflow Run` workflow in ic-private.