feat(sns): Add `RegisterExtension` proposal type (#5501)
This PR extends the SNS Governance API to support a new native proposal
type, `RegisterExtension`.

This type of proposals will be used for creating SNS-controlled
canisters that (unlike regular SNS dapp canisters) have certain
capabilities to operate on behalf of the DAO. In particular, the first
extension supported type will be for SNS treasury managers (not yet
implemented in this PR).

Deregistering an extension may be implemented via the
`DeregisterDappCanisters` proposal, or potentially via a new, dedicated
proposal; this design decision does not have to be made in this PR.

| [Next PR](https://github.com/dfinity/ic/pull/5587) >