chore: do a release_build on a workflow_call (#5588)
Make sure that when the `CI Main` workfow is called from the `Release
Testing` workflow `release_build` will be enabled to ensure the `Repro
check` job in the `Release Testing` workflow can download the expected
artifacts from the CDN.

Note that we implement this by enabling `release_build` when `'${{
github.event_name }}' == 'workflow_call'`. This means that when `CI
Main` is called from `Release Testing` artifacts will be uploaded.
However it also means that artifacts will be uploaded if `CI Main` is
called from *any other* workflows. We don't do that currently but it
could be a surprise when we do. A way to prevent this is to pass a
`caller_workflow: "Release Testing"` argument when calling `CI Main`
from `Release Testing` and check for that. However, I don't think it's
worth the complexity, but I'm happy to be convinced otherwise.