fix: Do not return system state changes on errors for composite queries (#5567)
This PR fixes what system state changes are returned in case an error is
encountered during execution of composite queries. In that case, no
changes should be returned.

The previous version of the code was returning some changes but since
the composite queries operate on a "throw-away" copy of the state which
is only visible to the composite query graph itself (i.e. it's isolated
from other composite query executions) and additionally in case of an
error, the error would be returned to the user immediately and not
continue executing any other downstream calls, there should have been no
impact really but it's still good to fix this to be consistent.