refactor: Move Guest VM management script to Rust (#5435)
This is a rewrite of the Guest VM management script (guestos.sh) to
Rust. The goal is to add typesafety and test common scenarios (e.g. VM
failing behavior).

The major important functional change is that configuration files are no
longer assumed to be generated by other services. Instead, the service
generates the necessary files before starting the VM which are cleaned
up when the process ends. This way we can reduce the coupling between
services and test related functionality together.

Some of the metrics are no longer meaningful and therefore have been
removed.

This change makes the generate-guestos-config service and related
implementation obsolete which will be cleaned up in a later PR.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: IDX GitHub Automation <>
Co-authored-by: Nicolas Mattia <nicolas.mattia@dfinity.org>