feat: Improvements for partition_tools (#5603)
This PR adds two new features:

- Allow opening a partition at a certain offset with certain length.
This is lower level than the existing index-based open.
- Allow extracting the entire file tree of a partition to a directory.
- There is also a small fix in the `check_offset` implementation which
no longer panics if the output of fdisk can't be parsed.

The new features will be used in the GuestVM qemu direct boot
implementation tests which includes extracting the contents of the GRUB
and boot partitions.