chore: add http outcalls cost metrics (#5562)
This PR is meant to compare a new suggested formula for the price of an
http request
[here](https://docs.google.com/document/d/1wcEJrDGmQGhZCrMcUxoDKKRe1IezqNUJIHsTolzXRTQ/edit?tab=t.0#heading=h.x75i679ukxrw).

For reference, the current formula described
[here](https://internetcomputer.org/docs/references/cycles-cost-formulas#https-outcalls):
```
N x (
  3_000_000 + 
  60_000 x N + 
  400 x request_bytes + 
  800 x max_response_bytes
)
```

is compared to the newly proposed:
```
N x (
  4_000_000 + 
  50_000 x N + 
  50 x request_bytes +
  50 x max_response_bytes + 
  750 x payload_size +  
  30 x N x payload_size
)
```

where `N` is the subnet size.

We measure histograms for each price, for their difference and for their
ratio.

In addition, we also build histograms for the individual relevant fields
(`request_bytes`, `max_response_bytes` and `payload_size`).