feat: proxy application subnet requests via the http gateway (#5597)
This makes application subnets outcalls have the following behaviour:
- Make the request directly
- If the direct request fails, use `socks5h://socks5.ic0.app:1080` as a
socks proxy and retry.

The socks proxy address currently resolves to API boundary nodes, which
will immediately reject the request as it's coming from an app subnet.

However in the future the domain name will resolve to some new
"application subnet dedicated" machines which will accept proxy requests
from nodes on the app subnet.

This is part of an effort to allow replicas to make requests to Stripe,
which only supports IPv4. Stripe will (currently) be the only allowed
domain name accessible, and that will be enforced on the proxy server.

This change will have no affect over:
- system subnets, as they no longer use the socks domain name, but
rather get the IPs from the registry
- successful outcalls on app subnets

Currently there are ~0.6 rps across all nodes and subnets that fail with
"connect", or in other words, the newly added codepath will be executed
0.6 times / s, which is ~1% of all http outcall requests.