refactor(nns): Refactor distribute_rewards to separate immutable and mutable operations (#5494)
# Why

The `distribute_rewards` method is quite large and it has a lot of
complicated calculations. When we look at the whole method, it's hard to
distinguish the calculations from the mutations (schedule rewards
distribution, process proposals etc.) For situations like this, it's
often better to separate the immutable operations from the mutable ones.

# What

* Refactor the calculation logic in `distribute_rewards` to
`calculate_voting_rewards`