refactor: [EXC-2052] Use separate System API types for response handlers in a composite query (#5561)
This is the follow-up to #5550 which introduces new System API types for
the composite query response handlers. This change allows us to cleanly
separate these contexts without the need to the execution mode and the
use of `if-else` statements within the previous reply or reject callback
contexts. After this change the implementation is following closely the
specification which also allows for easier inspection of correctness.