chore(IDX): Use simple action for fuzzers build
This changes the CI jobs for fuzzers to be built with the simple `bazel`
action instead of `bazel-test-all`. The fuzzers are only being built and
don't actually run `bazel test`.

This does mean that the `diff` script is not being used and the full
build will "run" every time, though in case of a cached build this only
involves downloading ~20 targets and the whole job takes less than 2
minutes to run.