chore(IDX): Simplify release-testing workflow
This changes the `release-testing.yml` workflow to use the simpler
`bazel` action, as the workflow does not require `bazel-test-all`'s bells and whistles. It also merges the system test jobs by merging them into one job, parametrized via build strategy. Finally, the jobs are changed to use the `stamped` config to ensure the actual to-be-released artifacts are used.