fix(test): Increase DKG interval of large subnet recovery test (#5649)
The NNS of the large subnet recovery test has 40 nodes. During recovery
we need to reshare 3 transcripts (high/local, high/remote, low/remote)
with 40 dealings each. Each data block holds at most 1 dealing. This
means the DKG interval length should be at least 120.

Additionally, we remove a check in the `ValidateReplayStep` returning an
error if there exists a node with a finalized height that is greater
than the replay height. This check shouldn't exist, as we only need to
ensure that we replay at least up to the latest certified height.