chore: [EXC-1997] Early parsing of `SnapshotId` (#5630)
This PR implements `De/Serialize` and `CandidType` for `SnapshotId`.
This enables us to use `SnapshotId` type directly in management canister
types (without specifying `#[with(serde_bytes)] Vec<u8>`), and the
de/serialization is delegated to the new impls, including the fallible
direction `Vec<u8> -> SnapshotId`. This has the added benefit that we no
longer use fields that are called `snapshot_id`, but contain a raw
`Vec<u8>`, and then needing to convert and unwrap throughout.

Furthermore, the PR moves the `try_from` logic of `SnapshotId` from the
trait impl to the struct's impl. The reason is that the generic bound
`AsRef<[u8]>` cannot be used in the trait impl because it would overlap
with a blanket impl in `core`.