feat: Wrap components as prep for container upgrade (#5604)
1. We depend on the system `strip` when compiling or linking. Updating
this binary in https://github.com/dfinity/ic/pull/5615 has an effect on
outputs that was not seen in the build graph. This adds a dependency on
the system `strip` binary to the `hermetic_cc` linker and compiler.
2. `podman export` behaves differently after
https://github.com/dfinity/ic/pull/5615. This adds an additional layer
of sanitizing so that builds before and after the upgrade are
consistent.