feat(ic-admin): Use canister_ranges_ keys to read routing table (#5616)
# Why

We are switching from a monolithic routing table to multiple smaller
registry entries. This PR changes ic-admin from reading to the old value
to the new ones.

# What

* Use `NnsDataProvider` to poll registry entries, select the keys
through `get_key_family`, and read the values based on the keys.

# Tested

The `ic-admin get-routing-table` yields the exact same result before and
after the change.