refactor(orchestrator): [CON-1457] Remove dependency to canister_client in orchestrator (#5686)
This PR reverts #5224, which reverted the original PR that removed the
dependency to `canister_client` in the orchestrator (#5104), thus
remerging the original feature.

The reason why it was reverted in the first place was because the
`//rs/tests/nested:registration` system test was using an API boundary
node that served a self-signed TLS certificate which was accepted by
`canister_client` but not by `ic-agent`. Thanks to #5675, the test now
uses a HTTP gateway that has a valid certificate, which both allows the
orchestrator to use `ic-agent` and also better reflects the mainnet
behaviour.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>