feat: Support SEV config and direct boot in the VM runner (#5614)
This PR adds support for QEMU direct boot and adds SEV config options to
the VM XML config when SEV is enabled in the HostOSConfig. QEMU direct
boot is enabled regardless of the config.

Since direct boot will skip the GRUB bootloader, the logic from the
existing GuestOS grub.cfg was moved into the HostOS VM runner
(`prepare_direct_boot` in `guest_direct_boot.rs`). The code mounts the
grub and boot partitions of the guest, updates the grubenv, extracts the
necessary files and kernel command line and returns these.

An interesting complication is that older GuestOS releases don't have
the boot_args file and so don't support direct boot. In this case we
need to fall back to booting with GRUB. In this case, we must not update
the grubenv, since the GRUB bootloader will update it again and it's a
not idempotent.

To support testing, there is a `mount` module which contains
abstractions for mounting and mocking the contents of the GuestOS disk.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>