fix: Remove readonly from guest_direct_boot mounts (#5699)
Mounting in readonly mode can fail if the filesystem needs to be
repaired or if the OS needs to write the partition for multi mount
protection. Since the implementation ensures that the mount is cleaned
up before the Guest VM is started, we don't really need to mount the
guestos device in readonly mode.

Also, add more context messages for easier debugging.