[feat]: Rehash manifest after publishing it [MR-682] (#5500)
Currently, we rehash 10% of the data to detect divergence in case of
e.g. system or hardware bug. In which case we expect to crash and state
sync after 10 checkpoint intervals on average.
This change moves out this rehashing check after the publishing of
manifest, increasing the average detection to 11 checkpoint intervals
but speeding up publishing the manifest to consensus.