feat(sns): add recently executed proposals to the sns metrics (#5744)
This PR adds the number of recently executed proposals to the SNS
metrics (see [this PR](https://github.com/dfinity/ic/pull/5079)). Please
note that when a proposal gets executed, it can later get garbage
collected and removed from the list of proposals by the governance.
Tracking these garbage collected proposals is not implemented in this
PR.