feat(nns): Use a previous snapshot for ballots if an unusually high voting power is detected (#4798)
# Why

To prevent a sudden takeover of the NNS in the event of some party
gaining illegitimate voting power (e.g. a bug in Governance, ICP Ledger,
etc.), we aim at disabling the early adoption of NNS proposals if the
voting power has a spike, which should give the network more time to
react and recover in such cases.

# What

There are 3 PRs to achieve the above-mentioned goal:

- (#4404) Define VotingPowerSnapshot and a collection of snapshots
- (#4405) Add a timer task to record snapshots periodically
- (current PR) Use the previous snapshot if a voting power spike is
detected