fix: ignore canister history memory usage in canister invariant checks (#5758)
This PR makes canister history memory usage ignored in canister
invariant checks. The motivation for this change is that the memory
usage is not enforced to stay within the memory allocation if canister
history size increases. Hence, the current check is too strict and could
lead to spurious critical errors.