refactor(cketh): clean up eth_rpc.rs (#5717)
[XC-224](https://dfinity.atlassian.net/browse/XC-224?atlOrigin=eyJpIjoiOTdjOGU5OWFlNzMwNGYxMWJiYzhhZmM0Y2NlZGQxODgiLCJwIjoiaiJ9):
This PR is the last in the series of PR aimed at removing legacy code
that enabled the ckETH minter to interact with Ethereum via direct HTTPS
outcalls (see https://github.com/dfinity/ic/pull/4926).

This PR performs further cleaning in the `eth_rpc.rs` file:
- Remove unused `Quantity` struct and related test
- Remove unused `Data` struct and related test
- Remove unused `MAX_PAYLOAD_SIZE` and `HTTP_MAX_SIZE`
- Move `HEADER_SIZE_LIMIT` to improve locality

The `MultiCallResults` struct was thought to be redundant and initially
envisioned to be removed. It turns out to be useful to perform `map` and
`reduce` which are not provided by the EVM RPC canister, in particular
in the following cases:
- `eth_fee_history`: to reduce with strict majority by key.
- `eth_send_raw_transaction`: to reduce with any of `Ok` result.
- `eth_get_latest_transaction_count`: to reduce with min by key.

[XC-224]:
https://dfinity.atlassian.net/browse/XC-224?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>