feat: enable non replicated outcalls (#5715)
This PR implements http requests with [`is_replicated =
Some(false)`](https://github.com/dfinity/ic/blob/6b57b14995fc16b483b8414d6c2dd537197bf1de/rs/types/management_canister_types/src/http.rs#L84).

If `is_replicated == Some(false)`, then a context with `replication =
NonReplicated(node_id)` is added to the replicated state. The node_id is
chosen by the execution environment randomly, consistently across all
replicas.

The client processor ignores such a request in the replicated state
unless its own `node_id` is equal to the one in
`NonReplicated(node_id)`. If that is the case, it makes the request
normally, as before, and gossips the metadata and signature (gossiping
this is not required, but it would've been more difficult to not do it;
also in the future we plan to gossip it anyway, and there is no harm in
gossiping it right now, so we keep it).

During payload building, the block maker checks its replicated state
agains http responses + shares in its pool. If an http request has
`NonReplicated(node_id)` replication, the block maker will require at
least one of the shares to originate from `node_id`, and to have the
full response in it's own pool as well. (for now, this also means that
its own node id is equal to `node_id`, however with gossiping, that may
not be necessarily true in the future). It will then include only the
relevant share in the payload.

During payload validation, the validator checks its replicated state
agains the http responses + shares in the payload. If a response
corresponds to a request which has `NonReplicated(node_id)`, then it
requires the payload to have _exactly_ one share, and that share should
corresponds to the delegated node_id (essentially lowering the signature
threshold from 2f + 1 to 1, and making the "committee" be represented by
just node_id).

The PR also includes one metric (an int counter), counting single
replica responses (which includes replicated responses if the subnet
only has one node).

The change was tested via a system test, by calling out to a `/random`
endpoint on a 4 node subnet.

<more than half of the added lines are tests; this is not actually a
huge change.>

Things to be on the lookout for:
1. Is a Malicious replica able to create a payload containing the
response to a fully replicated request, claiming it’s non replicated
(and thus only containing one signature)
2. Can a malicious replica dos the feature by broadcasting “valid”
shares? A “valid” share in this context is a share to a non replicated
http request which was not delegated to it. However, it’s valid from any
other standpoint (same content hash etc).
3. Can a malicious replica hijack the request and include it in the
block even if it was not delegated to it?
4. What happens if the node leaves the subnet? (request should timeout).
5. ??