fix(tests): Temporarily revert chain key performance tests back to `MessageCanister` (#5786)
https://github.com/dfinity/ic/pull/5530 introduced the `SignerCanister`
allowing us to generate signature requests as part of the canister
itself, instead of sending them as an ingress message.

However, responses of the `SignerCanister` are currently always
interpreted as successful, even if the actual signature request failed
(i.e., because the request queue was full). This is skewing the results
of the benchmark (see
[here](https://grafana.testnet.dfinity.network/d/crypto-benchmarks-criterion-tecdsa/tecdsa?from=now-30d&to=now&var-datasource=es-logs-ci-performance)
for instance). Therefore, we temporarily revert to the
`MessageCanister`.