chore(nns): Rename a metric according to convention (#5781)
Metrics in NNS Governance should begin with `governance_` according to
convention. This is a rather new metric (added less than one month ago),
so there shouldn't be dependencies on it. In addition, this is a boolean
metric indicating something goes wrong, so it's extremely unlikely that
some clients started to depend on it.