test(ICP_Rosetta): FI-1779: Ignore spender in transfers in search_transactions_by_account (#5794)
Ignore the `spender` account in `transfer_from` and `burn` operations
when searching for transactions by account in Rosetta. This is because
the account balance of the `spender` is not affected in a
`transfer_from` and `burn`, so no Rosetta operation is created for the
spender. In this case, only `approve` transactions will be found when
searching for the `spender` account.