feat(registry): Added `canister_cycles_cost_schedule` to CreateSubnetPayload. (#5777)
This doesn't actually do anything yet. I mean, callers can populate that
field, but if they set the field to `Some(Free)`, they will not get the
behavior indicated by `Free`. This is ok, because Governance is the only
principal that is allowed to call `create_canister`, and Governance does
not try to use `Free` (yet), but that will come in the not too distant
future.

# Future Work

The next step is to add a similar field to `SubnetRecord`, which would
then be populated by [SubnetRecord::from(create_subnet_payload)][from].

[from]:
https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/registry/canister/src/mutations/do_create_subnet.rs?L483-485