fix: EXC: Fix `memory_bytes` The PR fixes the `memory_bytes` getters around the codebase.
The implementation takes care of determinism
(by using deterministic `len`) and performance
(by not calculating the deep `Vec` or `BTreeSet`
sizes).

Due to the above, the size calculated by the
`memory_bytes` is still an estimation, not
an exact value.

The PR also adds a few tests to future-proof
the size calculations in query cache.