feat: Build custom SEV-compatible OVMF from source and distribute it with the GuestOS (#5776)
The custom OVMF is necessary to support SEV-SNP. Since the OVMF is used
for the calculation of the SEV measurement, we include it in the Guest.
This has the benefit that we get a consistent measurement for each
GuestOS release and we don't have to keep track of all OVMFs that may be
lying around in HostOS-es.

For backwards compatibility, we fall back to the OVMF taken from the
HostOS when the GuestOS has no OVMF.

---------

Co-authored-by: IDX GitHub Automation <>