feat(ICP-Ledger): FI-1550: Implement endpoint to retrieve approvals (#5680)
The endpoint returns the approvals for the given `AccountIdentifier`.
Since this information is also stored in the blockchain, the endpoint
does not check if the caller is the owner of the given
`AccountIdentifier`. The endpoint returns a maximum of 500 allowances at
a time, the `prev_spender_id` argument can be used for pagination. The
`take` argument can be used to additionally limit the number of returned
approvals.