feat: Added `canister_cycles_cost_schedule` to `SubnetRecord`. (#5816)
Like how we did with `CreateSubnetPayload` in [PR 5777].

The primary changes are in `subnet.proto`, and `do_create_subnet.rs`, so
look at those two files to see what's "really" going on here. Then, the
rest of the changes in this PR are just "corollary" changes from the
primary ones.

Most of the other changes are just adding

```
canister_cycles_cost_schedule: Normal as i32,
```

to existing instantiations of `SubnetRecord`, which mostly (entirely?)
occur in tests.

[PR 5777]: https://github.com/dfinity/ic/pull/5777

# References

Closes [NNS1-3924].

[NNS1-3924]: https://dfinity.atlassian.net/browse/NNS1-3924

[👈 Previous PR][PR 5777]