feat: bootstrap CMC on PocketIC instances (#5831)
The scope of this PR is bootstrapping the cycles minting canister (CMC)
when creating a PocketIC instance and keeping its content in sync with
the PocketIC topology (e.g., the set of application subnets when a new
subnet is created on an existing PocketIC instance).

The registry canister (WASM) is obtained from the bazel target
`@mainnet_cycles-minting-canister//file`.