chore: Flush state manager more often in state machine tests
With more asynchronous components in the state manager, tests tend to
trip up as they might not expect it. This small change makes sure that
when a test waits for the latest manifest, it also finishes all other
asynchronous operations. This should help with some flakiness in tests
when state machine tests are restarted or their directories reused.