test: Extend `create_subnet_test` to create multiple subnets at the same time (#5844)
The existing test only covered the creation of a single subnet. However
in practice, multiple create subnet proposals may be adopted at the same
time.

Furthermore, after the subnets were created, we now also check the
health of the NNS (in addition to the new subnets), to make sure it
didn't run into any problems.