fix: Fix the reboot behavior of GuestVM
With the "restart" libvirt setting, the VM is restarted using the same direct boot args as it was used to start with. This is obviously wrong and breaks upgrades where changing the boot params is necessary.

This fix doesn't allow QEMU to restart the VM but rather adds support in the guest_vm_runner to recreate the VM with a brand new configuration.