feat: CON-1527 Create a collection of pre-signatures in replicated state (#5768)
Currently, pre-signatures required to answer tECDSA and tSchnorr
requests are stored in the block payload. This greatly limits the number
of pre-signatures we can compute in advance. In the future, we want to
deliver and store available pre-signatures in the replicated state
instead.

This PR creates the collection in replicated state which will be holding
delivered pre-signatures. In addition to the pre-signatures, each
"pre-signature stash" also holds the current key transcript. In the
future, whenever the key transcript changes (i.e. due to a key rotation
or membership change), the corresponding pre-signatures shall be purged.

For now, the new collection will remain unused for the foreseeable
future, which is necessary to ensure replica version compatibility.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>